+++ css/csstools-design-tokens/src/extension.ts
@@ -1,15 +1,33 @@
 import * as vscode from 'vscode';
 import postcss from 'postcss';
-import { PluginContainer, pluginContainer } from './postcss/index';
+import { PluginContainer, pluginContainer, pluginOptions } from './postcss/index';
 
 const log = vscode.window.createOutputChannel("CSSTools Design Tokens");
+
+function getOptions(): pluginOptions {
+	let importAtRuleName = vscode.workspace.getConfiguration('designTokens')?.get('importAtRuleName') ?? 'design-tokens';
+	let valueFunctionName = vscode.workspace.getConfiguration('designTokens')?.get('valueFunctionName') ?? 'design-token';
+
+	if (!(typeof importAtRuleName === 'string') || !importAtRuleName) {
+		importAtRuleName = 'design-tokens';
+	}
+
+	if (!(typeof valueFunctionName === 'string') || !valueFunctionName) {
+		valueFunctionName = 'design-token';
+	}
+
+	return {
+		importAtRuleName: importAtRuleName as string,
+		valueFunctionName: valueFunctionName as string,
+	};
+}
 
 export function activate(context: vscode.ExtensionContext) {
 	const tokensData: Map<string, PluginContainer> = new Map();
 
 	{
 		if (vscode.window?.activeTextEditor?.document?.languageId === 'css') {
-			handleCSSDocument(tokensData, vscode.window.activeTextEditor.document);
+			handleCSSDocument(tokensData, vscode.window.activeTextEditor.document, getOptions());
 		}
 	}
 
@@ -18,7 +36,7 @@ export function activate(context: vscode.ExtensionContext) {
 			return;
 		}
 
-		handleCSSDocument(tokensData, document);
+		handleCSSDocument(tokensData, document, getOptions());
 	}));
 
 	context.subscriptions.push(vscode.window.onDidChangeActiveTextEditor((editor) => {
@@ -30,40 +48,48 @@ export function activate(context: vscode.ExtensionContext) {
 			return;
 		}
 
-		handleCSSDocument(tokensData, editor.document);
+		handleCSSDocument(tokensData, editor.document, getOptions());
 	}));
 
 	context.subscriptions.push(vscode.languages.registerCompletionItemProvider('css', {
 		async provideCompletionItems(document, position, cancelToken) {
-			const tokens = tokensData.get(document.uri.fsPath);
-			if (!tokens) {
-				return;
-			}
+			const options = getOptions();
 
 			const completionItems: Array<vscode.CompletionItem> = [];
 
-			if (position.character === 0) {
-				const completionItem = new vscode.CompletionItem('@design-tokens', vscode.CompletionItemKind.Snippet);
-				completionItem.insertText = new vscode.SnippetString(
-					"@design-tokens url('$1') format('${2:style-dictionary3}');$0"
-				);
-				completionItem.detail = 'Import design tokens';
-				completionItems.push(completionItem);
+			log.appendLine(document.lineAt(position.line).text);
+
+			if (position.character === 0 || document.lineAt(position.line).text.startsWith('@' + options.importAtRuleName)) {
+				{
+					const completionItem = new vscode.CompletionItem('@' + options.importAtRuleName, vscode.CompletionItemKind.Snippet);
+					completionItem.insertText = new vscode.SnippetString(
+						"@" + options.importAtRuleName + " url('$1') format('${2:style-dictionary3}');$0"
+					);
+					completionItem.filterText = '@' + options.importAtRuleName;
+					completionItem.detail = 'Import design tokens';
+					completionItems.push(completionItem);
+				}
+
+				{
+					const completionItem = new vscode.CompletionItem('@' + options.importAtRuleName + ' when', vscode.CompletionItemKind.Snippet);
+					completionItem.insertText = new vscode.SnippetString(
+						"@" + options.importAtRuleName + " url('$1') format('${2:style-dictionary3}') when('$3');$0"
+					);
+					completionItem.filterText = '@' + options.importAtRuleName + ' when';
+					completionItem.detail = 'Import design tokens conditionally';
+					completionItems.push(completionItem);
+				}
 			}
 
-			if (position.character === 0) {
-				const completionItem = new vscode.CompletionItem('@design-tokens when', vscode.CompletionItemKind.Snippet);
-				completionItem.insertText = new vscode.SnippetString(
-					"@design-tokens url('$1') format('${2:style-dictionary3}') when('$3');$0"
-				);
-				completionItem.detail = 'Import design tokens conditionally';
-				completionItems.push(completionItem);
+			const tokens = tokensData.get(document.uri.fsPath);
+			if (!tokens) {
+				return completionItems;
 			}
 
 			for (const [tokenPath, token] of tokens.tokens().entries()) {
 				const completionItem = new vscode.CompletionItem(tokenPath + ` ${token.cssValue()}`, vscode.CompletionItemKind.Variable);
-				completionItem.insertText = `design-token('${tokenPath}')`;
-				completionItem.filterText = `design-token('${tokenPath}') ${token.cssValue()}`;
+				completionItem.insertText = `${options.valueFunctionName}('${tokenPath}')`;
+				completionItem.filterText = `${options.valueFunctionName}('${tokenPath}') ${token.cssValue()}`;
 
 				if (token.comment) {
 					completionItem.documentation = new vscode.MarkdownString(token.comment);
@@ -123,9 +149,9 @@ export function activate(context: vscode.ExtensionContext) {
 // this method is called when your extension is deactivated
 export function deactivate() {}
 
-async function handleCSSDocument(tokensData: Map<string, PluginContainer>, document: vscode.TextDocument) {
+async function handleCSSDocument(tokensData: Map<string, PluginContainer>, document: vscode.TextDocument, options: pluginOptions) {
 	const css = document.getText();
-	const documentPlugin = tokensData.has(document.uri.fsPath) ? tokensData.get(document.uri.fsPath)! : (tokensData.set(document.uri.fsPath, pluginContainer()), tokensData.get(document.uri.fsPath)!);
+	const documentPlugin = tokensData.has(document.uri.fsPath) ? tokensData.get(document.uri.fsPath)! : (tokensData.set(document.uri.fsPath, pluginContainer(options)), tokensData.get(document.uri.fsPath)!);
 	documentPlugin.reset();
 
 	try {