+++ css/csstools-design-tokens/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## 1.1.0
+
+- Added `designTokens.valueFunctionName` option to control the design-token function name.
+- Added `designTokens.importAtRuleName` option to control the design-tokens at-rule name.
+
+```json
+  "designTokens.importAtRuleName": "gimme-tokens",
+  "designTokens.valueFunctionName": "token"
+```
+
 ## 1.0.2
 
 - Various improvements to completion item triggers.